# swagger/file.json
{"swagger":"2.0","info":{"description":"接口文档","version":"1.0","title":"Swagger2 接口文档","contact":{"name":"luofeng","url":"http://pingpingpang.cn","email":"982338665@qq.com"}},"host":"127.0.0.1:8080","basePath":"/","tags":[{"name":"Hello Controller测试","description":"Hello Controller"},{"name":"index-controller","description":"Index Controller"}],"paths":{"/test/1":{"get":{"tags":["Hello Controller测试"],"summary":"测试1","description":"根据ID获取用户信息详细描述","operationId":"sayUsingGET","produces":["*/*"],"parameters":[{"name":"Authorization","in":"header","description":"AccessToken令牌","required":false,"type":"string"},{"name":"name","in":"query","description":"name","required":false,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/test/2":{"get":{"tags":["Hello Controller测试"],"summary":"say2","operationId":"say2UsingGET","produces":["*/*"],"parameters":[{"name":"Authorization","in":"header","description":"AccessToken令牌","required":false,"type":"string"},{"name":"age","in":"query","required":false,"type":"integer","format":"int32"},{"name":"createTime","in":"query","required":false,"type":"string","format":"date-time"},{"name":"deleted","in":"query","required":false,"type":"integer","format":"int32"},{"name":"email","in":"query","required":false,"type":"string"},{"name":"id","in":"query","description":"主键ID","required":false,"type":"integer","format":"int64","x-example":1},{"name":"managerId","in":"query","required":false,"type":"integer","format":"int64"},{"name":"name","in":"query","description":"姓名","required":false,"type":"string","x-example":"张三"},{"name":"updateTime","in":"query","required":false,"type":"string","format":"date-time"},{"name":"version","in":"query","required":false,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/User"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/test/3":{"post":{"tags":["Hello Controller测试"],"summary":"say3","operationId":"say3UsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"Authorization","in":"header","description":"AccessToken令牌","required":false,"type":"string"},{"in":"body","name":"users","description":"users","required":true,"schema":{"$ref":"#/definitions/TestVo"}}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"type":"string"}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/test/4":{"post":{"tags":["Hello Controller测试"],"summary":"say4","operationId":"say4UsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"Authorization","in":"header","description":"AccessToken令牌","required":false,"type":"string"},{"in":"body","name":"users","description":"users","required":true,"schema":{"type":"array","items":{"$ref":"#/definitions/用户实体类"}}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/TestVo"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/test/{id}/{name}":{"delete":{"tags":["Hello Controller测试"],"summary":"删除用户","description":"删除用户详细描述","operationId":"deleteUsersUsingDELETE","produces":["*/*"],"parameters":[{"name":"Authorization","in":"header","description":"AccessToken令牌","required":false,"type":"string"},{"name":"id","in":"path","description":"用户的唯一标识","required":true,"type":"string"},{"name":"name","in":"path","description":"用户名","required":false,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"204":{"description":"No Content"},"401":{"description":"禁止访问"},"403":{"description":"Forbidden"}},"deprecated":false}},"/test1/test1":{"get":{"tags":["index-controller"],"summary":"say","operationId":"sayUsingGET_1","produces":["*/*"],"parameters":[{"name":"Authorization","in":"header","description":"AccessToken令牌","required":false,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}}},"definitions":{"Data":{"type":"object","title":"Data"},"TestVo":{"type":"object","properties":{"age":{"type":"integer","format":"int32"},"array":{"type":"array","items":{"type":"string"}},"money":{"type":"number","format":"double"},"testVo":{"$ref":"#/definitions/TestVo"},"testVos":{"type":"array","items":{"$ref":"#/definitions/TestVo"}},"time":{"$ref":"#/definitions/Data"},"users":{"$ref":"#/definitions/用户实体类"},"usersArray":{"type":"array","items":{"$ref":"#/definitions/用户实体类"}},"usersList":{"type":"array","items":{"$ref":"#/definitions/用户实体类"}}},"title":"TestVo"},"User":{"type":"object","properties":{"age":{"type":"integer","format":"int32"},"createTime":{"type":"string","format":"date-time"},"email":{"type":"string"},"id":{"type":"integer","format":"int64"},"managerId":{"type":"integer","format":"int64"},"name":{"type":"string"},"remark1":{"type":"string"},"remark3":{"type":"string"}},"title":"User"},"用户实体类":{"type":"object","properties":{"age":{"type":"integer","format":"int32"},"createTime":{"type":"string","format":"date-time"},"deleted":{"type":"integer","format":"int32"},"email":{"type":"string"},"id":{"type":"integer","format":"int64","example":1,"description":"主键ID"},"managerId":{"type":"integer","format":"int64"},"name":{"type":"string","example":"张三","description":"姓名"},"updateTime":{"type":"string","format":"date-time"},"version":{"type":"integer","format":"int32"}},"title":"用户实体类"}}}